Update 1_Expiration_Date_Verification_✔️.py
Set the constraint on the date so the checkbox and info message only would show on the last Friday, if the next Monday is going to be next month. When the checkbox is checked, the manufacturing date would be default set to next Monday, which has the expiration date the first day of coming month, three years later.